EllipseSampling

Documentation for EllipseSampling.

EllipseSampling.calculate_ellipse_parameters

EllipseSampling.x_parametric_equation

EllipseSampling.y_parametric_equation

EllipseSampling.calculate_ellipse_parameters — Method

calculate_ellipse_parameters(Γ::Matrix{Float64}, ind1::Int, ind2::Int,
    confidence_level::Float64)

Given a square matrix Γ, the inverse of the Hessian of a log-likelihood function at its maximum likelihood estimate, indexes of the two variables of interest and the confidence level to construct a 2D ellipse approximation of the log-likelihood function, return the parameters of that ellipse; a and b, the radius of the major and minor axis respectively, x_radius and y_radius, the radius of the ellipse in the x and y axis respectively (i.e. the radius when the rotation α is zero) and α, an angle between 0 and 2π radians that the ellipse has been rotated by. a is equal to the maximum of x_radius and y_radius, while b is equal to the minimum of x_radius and y_radius.

Arguments

Γ: A square matrix (2D) which is the inverse of the Hessian of a log-likelihood function at its maximum likelihood estimate.

ind1: Index of the first parameter of interest (corresponds to the row and column index of Γ)

ind2: Index of the second parameter of interest (corresponds to the row and column index of Γ).

confidence_level: The confidence level ∈[0.0,1.0] at which the ellipse approximation is constructed.

EllipseSampling.x_parametric_equation — Method

x_parametric_equation(t::T, x_radius::T, y_radius::T, α::T=0.0, Cx::T=0.0) where T<:Float64

Implements the parametric equation for variable x of a translated and rotated ellipse, $x(t)$, where t is an angle between 0 and 2π radians.

Arguments

t: an angle between 0 and 2π radians that defines the x location on the ellipse.

x_radius: radius of the ellipse in the x axis (i.e. when the rotation, α, is zero).

y_radius: radius of the ellipse in the y axis (i.e. when the rotation, α, is zero).

α: an angle between 0 and 2π radians that the ellipse has been rotated by.

Cx: the x coordinate of the centre of the ellipse (the translation of the ellipse in the x axis).

EllipseSampling.y_parametric_equation — Method

y_parametric_equation(t::T, x_radius::T, y_radius::T, α::T, Cx::T) where T<:Float64

Implements the parametric equation for variable y of a translated and rotated ellipse, $y(t)$, where t is an angle between 0 and 2π radians.

Arguments

t: an angle between 0 and 2π radians that defines the y location on the ellipse.

x_radius: radius of the ellipse in the x axis (i.e. when the rotation, α, is zero).

y_radius: radius of the ellipse in the y axis (i.e. when the rotation, α, is zero).

α: an angle between 0 and 2π radians that the ellipse has been rotated by.

Cy: the y coordinate of the centre of the ellipse (the translation of the ellipse in the y axis).