Internal Library

Documentation for structs and functions not covered within User Interface documentation.

Structs

EllipseSampling.Ellipse — Type

Ellipse(x_radius::Float64,
    y_radius::Float64,
    α::Float64,
    Cx::Float64,
    Cy::Float64,
    a::Float64,
    b::Float64,
    m::Float64,
    circumference::Float64)

Contains the information required to define an ellipse which may have been rotated and translated. See construct_ellipse

Arguments

x_radius: radius of the ellipse in the x axis (i.e. when the rotation, α, is zero).

y_radius: radius of the ellipse in the y axis (i.e. when the rotation, α, is zero).

α: an angle in radians (0 to 2π) that the ellipse has been rotated by. A positive value represents an anti-clockwise rotation.

Cx: the x coordinate of the centre of the ellipse (the translation of the ellipse in the x axis).

Cy: the y coordinate of the centre of the ellipse (the translation of the ellipse in the y axis).

a: the major radius of the ellipse.

b: the minor radius of the ellipse.

m: the eccentricity of the ellipse squared. See eccentricity_squared

circumference: the circumference of the ellipse.

Functions

EllipseSampling.eccentricity_squared — Function

eccentricity_squared(a::T, b::T) where T<:Float64

Computes m, the eccentricity of the ellipse squared, $m=e^2$, where $m=1-\big(\frac{b}{a}\big)^2$.

Arguments

a: the major radius of the ellipse.

b: the minor radius of the ellipse.

Details

This relationship between m and e is seen by considering the equation for e: $e=\frac{c}{a}$, where $c^2=\big(a^2-b^2\big)$ and $a>b$.

Replacing c in the equation for e:

\[e=\frac{\sqrt{a^2-b^2}}{a}\]

Substituting the equation for e into m:

\[m = \bigg(\frac{\sqrt{a^2-b^2}}{a}\bigg)^2 = \frac{a^2-b^2}{a^2} = 1 - \frac{b^2}{a^2} = 1 - \bigg(\frac{b}{a}\bigg)^2\]

EllipseSampling.circumference — Function

circumference(a::Float64, b::Float64)

Calculates the circumference of an ellipse using the elliptic integral of the second kind. Uses Elliptic.jl.

Arguments

a: the major radius of the ellipse.

b: the minor radius of the ellipse.

EllipseSampling.assert_parameters_are_valid — Function

assert_parameters_are_valid(a::T, b::T, x_radius::T, y_radius::T) where T<:Float64

Asserts that the parameters relate to a valid ellipse. I.e. that a ≥ b and, x_radius and y_radius are positive. Note: a=max(x_radius, y_radius), b=min(x_radius, y_radius).

Arguments

a: the major radius of the ellipse.

b: the minor radius of the ellipse.

x_radius: radius of the ellipse in the x axis (i.e. when the rotation, α, is zero).

y_radius: radius of the ellipse in the y axis (i.e. when the rotation, α, is zero).

Missing docstring.

Missing docstring for EllipseSampling.E_inverse. Check Documenter's build log for details.

EllipseSampling.calculate_ellipse_parameters — Function

calculate_ellipse_parameters(Γ::Matrix{Float64}, ind1::Int, ind2::Int,
    confidence_level::Float64)

Given a square matrix Γ, the inverse of the Hessian of a log-likelihood function at its maximum likelihood estimate, indexes of the two variables of interest and the confidence level to construct a 2D ellipse approximation of the log-likelihood function, return the parameters of that ellipse.

Arguments

Γ: A square matrix (2D) which is the inverse of the Hessian of a log-likelihood function at its maximum likelihood estimate.

ind1: Index of the first parameter of interest (corresponds to the row and column index of Γ)

ind2: Index of the second parameter of interest (corresponds to the row and column index of Γ).

confidence_level: The confidence level ∈[0.0,1.0] at which the ellipse approximation is constructed.

Details

The parameters of interest are a and b, the radius of the major and minor axis respectively, x_radius and y_radius, the radius of the ellipse in the x and y axis respectively (i.e. the radius when the rotation α is zero) and α, an angle between 0 and 2π radians that the ellipse has been rotated by. a is equal to the maximum of x_radius and y_radius, while b is equal to the minimum of x_radius and y_radius.

References and equation(s) to come.